SGBD/DBMS Vuelos Aereos
This program simulates the behavior of a DBMS... It is a work in progress intended for the class Data Structure for the Universidad Autonoma de Santo Domingo UASD university.

The program gathers information from the system catalog (the 4 main csv files : SysTable, SysColumn, SysForeignKey, SysPrimaryKey) and it creates the Table(Entity) in memory using the SysEntity Class.

When the program is finished it will do the basic operations of a DBMS to maintain data integrity, checks for valid data and so much more.

Before getting your hands into the code i suggest you have a look at the pdf proyectofinal so you will have a better understanding on what is being done.

I'll keep udating this program when i have more time.